fix: remove transparent border around resized images
