+++ backend/api/src/image_ops.rs
@@ -1,24 +1,33 @@
-use image::{imageops, DynamicImage, GenericImageView, RgbaImage};
-use imageops::FilterType;
+use image::{imageops::FilterType, DynamicImage, ImageOutputFormat};
 use shared::domain::image::ImageKind;
 
-pub fn generate_thumbnail(original: &DynamicImage) -> RgbaImage {
-    let (width, height) = ImageKind::THUMBNAIL_SIZE;
-    let thumb = original.thumbnail(width, height);
-    overlay_centered(&thumb, width, height)
-}
+pub fn generate_images(
+    original: DynamicImage,
+    kind: ImageKind,
+) -> anyhow::Result<(Vec<u8>, Vec<u8>, Vec<u8>)> {
+    let resized = {
+        let mut buffer = Vec::new();
+        let (width, height) = kind.size();
+        original
+            .resize(width, height, FilterType::Nearest)
+            .write_to(&mut buffer, ImageOutputFormat::Png)?;
+        buffer
+    };
 
-pub fn generate_resized(original: &DynamicImage, kind: ImageKind) -> RgbaImage {
-    let (width, height) = kind.size();
-    let resized = original.resize(width, height, FilterType::Nearest);
-    overlay_centered(&resized, width, height)
-}
+    let thumbnail = {
+        let mut buffer = Vec::new();
+        let (width, height) = ImageKind::THUMBNAIL_SIZE;
+        original
+            .thumbnail(width, height)
+            .write_to(&mut buffer, ImageOutputFormat::Png)?;
+        buffer
+    };
 
-fn overlay_centered(image: &DynamicImage, width: u32, height: u32) -> RgbaImage {
-    let (x, y) = ((image.width() - width) / 2, (image.height() - height) / 2);
+    let original = {
+        let mut buffer = Vec::new();
+        original.write_to(&mut buffer, ImageOutputFormat::Png)?;
+        buffer
+    };
 
-    let mut background = RgbaImage::new(width, height);
-
-    imageops::overlay(&mut background, image, x, y);
-    background
+    Ok((original, resized, thumbnail))
 }
